requirements.txt
# Core LLM Framework
litellm>=1.0.0

# Data validation
pydantic>=2.0.0

# Web scraping
beautifulsoup4>=4.14.0
feedparser>=6.0.10
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing (optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0

scikit-learn>=1.3.0
numpy>=1.25.0
pandas>=2.1.0
plotly>=5.15.0

datasets>=2.4.0
transformers>=4.57.3
sentence_transformers==5.2.0
peft>=0.14.0
bitsandbytes>=0.39.0
modal>=1.14.0
torch>=2.9.0

# Pushover for notifications
python-pushover>=0.3.0
# OpenAI API
openai>=2.14.0
litellm>=1.80.11

langraph>=0.5.0

chromadb>=1.4.0
bs4>=0.0.2
feedparser>=6.0.12
gradio>=6.2.0
plotly>=6.5.0
datasets>=4.4.2